+++ client/src/pages/Lessons/LessonsContentHeader.tsx
