+++ client/src/pages/Students/Students.tsx
@@ -1,0 +1,3 @@
+export function StudentsPage() {
+  return <></>;
+}
