+++ client/src/pages/Students/index.tsx
@@ -1,0 +1,1 @@
+export { StudentsPage } from './Students';
